Attempted fix for font

+++ GameProject/Project1/Source/Utilities.cpp
@@ -41,7 +41,7 @@ namespace utilities {
 	}
 
 	float getAngle(float x1, float y1, float x2, float y2) {
-		return AEACos((x1 - x2) / sqrt(((x1 - x2) * (x1 - x2)) + ((y1 - y2) * (y1 - y2))));
+		return AEACos((x1 - x2) / (float)sqrt(((x1 - x2) * (x1 - x2)) + ((y1 - y2) * (y1 - y2))));
 	}
